+++ plugins/system/Direct3D11VideoDevice.cpp
@@ -548,7 +548,7 @@ namespace Gek
             if (object)
             {
                 StringUTF8 nameUTF8(name);
-                object->SetPrivateData(WKPDID_D3DDebugObjectName, nameUTF8.length(), nameUTF8.c_str());
+                object->SetPrivateData(WKPDID_D3DDebugObjectName, UINT(nameUTF8.length()), nameUTF8.c_str());
             }
         }
 
@@ -566,7 +566,7 @@ namespace Gek
             template <typename CONVERT, typename INPUT>
             void set(const std::vector<INPUT> &inputList)
             {
-                uint32_t listCount = inputList.size();
+                size_t listCount = inputList.size();
                 objectList.reserve(std::max(listCount, objectList.size()));
                 objectList.resize(listCount);
 
@@ -576,7 +576,7 @@ namespace Gek
                 }
             }
 
-            void clear(uint32_t listCount)
+            void clear(size_t listCount)
             {
                 objectList.reserve(std::max(listCount, objectList.size()));
                 objectList.resize(listCount);
@@ -930,7 +930,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         samplerStateCache.set<SamplerState>(list);
-                        d3dDeviceContext->CSSetSamplers(firstStage, list.size(), samplerStateCache.get());
+                        d3dDeviceContext->CSSetSamplers(firstStage, UINT(list.size()), samplerStateCache.get());
                     }
 
                     ObjectCache<ID3D11Buffer> constantBufferCache;
@@ -939,7 +939,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         constantBufferCache.set<Buffer>(list);
-                        d3dDeviceContext->CSSetConstantBuffers(firstStage, list.size(), constantBufferCache.get());
+                        d3dDeviceContext->CSSetConstantBuffers(firstStage, UINT(list.size()), constantBufferCache.get());
                     }
 
                     ObjectCache<ID3D11ShaderResourceView> resourceCache;
@@ -948,7 +948,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         resourceCache.set<ShaderResourceView>(list);
-                        d3dDeviceContext->CSSetShaderResources(firstStage, list.size(), resourceCache.get());
+                        d3dDeviceContext->CSSetShaderResources(firstStage, UINT(list.size()), resourceCache.get());
                     }
 
                     ObjectCache<ID3D11UnorderedAccessView> unorderedAccessCache;
@@ -957,7 +957,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         unorderedAccessCache.set<UnorderedAccessView>(list);
-                        d3dDeviceContext->CSSetUnorderedAccessViews(firstStage, list.size(), unorderedAccessCache.get(), countList);
+                        d3dDeviceContext->CSSetUnorderedAccessViews(firstStage, UINT(list.size()), unorderedAccessCache.get(), countList);
                     }
 
                     void clearSamplerStateList(uint32_t count, uint32_t firstStage)
@@ -1025,7 +1025,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         samplerStateCache.set<SamplerState>(list);
-                        d3dDeviceContext->VSSetSamplers(firstStage, list.size(), samplerStateCache.get());
+                        d3dDeviceContext->VSSetSamplers(firstStage, UINT(list.size()), samplerStateCache.get());
                     }
 
                     ObjectCache<ID3D11Buffer> constantBufferCache;
@@ -1034,7 +1034,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         constantBufferCache.set<Buffer>(list);
-                        d3dDeviceContext->VSSetConstantBuffers(firstStage, list.size(), constantBufferCache.get());
+                        d3dDeviceContext->VSSetConstantBuffers(firstStage, UINT(list.size()), constantBufferCache.get());
                     }
 
                     ObjectCache<ID3D11ShaderResourceView> resourceCache;
@@ -1043,7 +1043,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         resourceCache.set<ShaderResourceView>(list);
-                        d3dDeviceContext->VSSetShaderResources(firstStage, list.size(), resourceCache.get());
+                        d3dDeviceContext->VSSetShaderResources(firstStage, UINT(list.size()), resourceCache.get());
                     }
 
                     void setUnorderedAccessList(const std::vector<Video::Object *> &list, uint32_t firstStage, uint32_t *countList)
@@ -1113,7 +1113,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         samplerStateCache.set<SamplerState>(list);
-                        d3dDeviceContext->GSSetSamplers(firstStage, list.size(), samplerStateCache.get());
+                        d3dDeviceContext->GSSetSamplers(firstStage, UINT(list.size()), samplerStateCache.get());
                     }
 
                     ObjectCache<ID3D11Buffer> constantBufferCache;
@@ -1122,7 +1122,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         constantBufferCache.set<Buffer>(list);
-                        d3dDeviceContext->GSSetConstantBuffers(firstStage, list.size(), constantBufferCache.get());
+                        d3dDeviceContext->GSSetConstantBuffers(firstStage, UINT(list.size()), constantBufferCache.get());
                     }
 
                     ObjectCache<ID3D11ShaderResourceView> resourceCache;
@@ -1131,7 +1131,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         resourceCache.set<ShaderResourceView>(list);
-                        d3dDeviceContext->GSSetShaderResources(firstStage, list.size(), resourceCache.get());
+                        d3dDeviceContext->GSSetShaderResources(firstStage, UINT(list.size()), resourceCache.get());
                     }
 
                     void setUnorderedAccessList(const std::vector<Video::Object *> &list, uint32_t firstStage, uint32_t *countList)
@@ -1201,7 +1201,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         samplerStateCache.set<SamplerState>(list);
-                        d3dDeviceContext->PSSetSamplers(firstStage, list.size(), samplerStateCache.get());
+                        d3dDeviceContext->PSSetSamplers(firstStage, UINT(list.size()), samplerStateCache.get());
                     }
 
                     ObjectCache<ID3D11Buffer> constantBufferCache;
@@ -1210,7 +1210,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         constantBufferCache.set<Buffer>(list);
-                        d3dDeviceContext->PSSetConstantBuffers(firstStage, list.size(), constantBufferCache.get());
+                        d3dDeviceContext->PSSetConstantBuffers(firstStage, UINT(list.size()), constantBufferCache.get());
                     }
 
                     ObjectCache<ID3D11ShaderResourceView> resourceCache;
@@ -1219,7 +1219,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         resourceCache.set<ShaderResourceView>(list);
-                        d3dDeviceContext->PSSetShaderResources(firstStage, list.size(), resourceCache.get());
+                        d3dDeviceContext->PSSetShaderResources(firstStage, UINT(list.size()), resourceCache.get());
                     }
 
                     ObjectCache<ID3D11UnorderedAccessView> unorderedAccessCache;
@@ -1228,7 +1228,7 @@ namespace Gek
                         GEK_REQUIRE(d3dDeviceContext);
 
                         unorderedAccessCache.set<UnorderedAccessView>(list);
-                        d3dDeviceContext->OMSetRenderTargetsAndUnorderedAccessViews(D3D11_KEEP_RENDER_TARGETS_AND_DEPTH_STENCIL, nullptr, nullptr, firstStage, list.size(), unorderedAccessCache.get(), countList);
+                        d3dDeviceContext->OMSetRenderTargetsAndUnorderedAccessViews(D3D11_KEEP_RENDER_TARGETS_AND_DEPTH_STENCIL, nullptr, nullptr, firstStage, UINT(list.size()), unorderedAccessCache.get(), countList);
                     }
 
                     void clearSamplerStateList(uint32_t count, uint32_t firstStage)
@@ -1334,14 +1334,14 @@ namespace Gek
                 {
                     GEK_REQUIRE(d3dDeviceContext);
 
-                    d3dDeviceContext->RSSetViewports(viewPortList.size(), (D3D11_VIEWPORT *)viewPortList.data());
+                    d3dDeviceContext->RSSetViewports(UINT(viewPortList.size()), (D3D11_VIEWPORT *)viewPortList.data());
                 }
 
                 void setScissorList(const std::vector<Video::ScissorBox> &rectangleList)
                 {
                     GEK_REQUIRE(d3dDeviceContext);
 
-                    d3dDeviceContext->RSSetScissorRects(rectangleList.size(), (D3D11_RECT *)rectangleList.data());
+                    d3dDeviceContext->RSSetScissorRects(UINT(rectangleList.size()), (D3D11_RECT *)rectangleList.data());
                 }
 
                 void clearResource(Video::Object *object, const Math::Float4 &value)
@@ -1416,7 +1416,7 @@ namespace Gek
                     GEK_REQUIRE(d3dDeviceContext);
 
                     renderTargetViewCache.set<RenderTargetView>(renderTargetList);
-                    d3dDeviceContext->OMSetRenderTargets(renderTargetList.size(), renderTargetViewCache.get(), getObject<DepthTexture>(depthBuffer));
+                    d3dDeviceContext->OMSetRenderTargets(UINT(renderTargetList.size()), renderTargetViewCache.get(), getObject<DepthTexture>(depthBuffer));
                 }
 
                 void setRenderState(Video::Object *renderState)
@@ -1466,7 +1466,7 @@ namespace Gek
                 {
                     GEK_REQUIRE(d3dDeviceContext);
                     
-                    uint32_t vertexBufferCount = vertexBufferList.size();
+                    uint32_t vertexBufferCount = UINT(vertexBufferList.size());
                     vertexBufferStrideCache.resize(vertexBufferCount);
                     vertexBufferOffsetsCache.resize(vertexBufferCount);
                     for (uint32_t buffer = 0; buffer < vertexBufferCount; buffer++)
@@ -2275,7 +2275,7 @@ namespace Gek
                 }
 
                 CComPtr<ID3D11InputLayout> d3dInputLayout;
-                HRESULT resultValue = d3dDevice->CreateInputLayout(d3dElementList.data(), d3dElementList.size(), compiledData, compiledSize, &d3dInputLayout);
+                HRESULT resultValue = d3dDevice->CreateInputLayout(d3dElementList.data(), UINT(d3dElementList.size()), compiledData, compiledSize, &d3dInputLayout);
                 if (FAILED(resultValue) || !d3dInputLayout)
                 {
                     throw Video::CreateObjectFailed();
@@ -2626,7 +2626,7 @@ namespace Gek
                     throw Video::CreateObjectFailed();
                 }
 
-                return std::make_shared<ViewTexture>(d3dResource.p, d3dShaderResourceView.p, nullptr, Video::Format::Unknown, image.GetMetadata().width, image.GetMetadata().height, image.GetMetadata().depth);
+                return std::make_shared<ViewTexture>(d3dResource.p, d3dShaderResourceView.p, nullptr, Video::Format::Unknown, uint32_t(image.GetMetadata().width), uint32_t(image.GetMetadata().height), uint32_t(image.GetMetadata().depth));
             }
 
             void executeCommandList(Video::Object *commandList)
